pyproject.toml
[project]
name = "FabricaNeeds-Back-Equipe5"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "django>=5.0.4",
    "djangorestframework>=3.15.1",
    "dotenv>=0.0.5",
    "python-dotenv>=1.0.1",
    "psycopg2>=2.9.9",
    "psycopg>=3.1.18",
    "psycopg2-binary>=2.9.9",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
